requirements.txt
selenium
webdriver_manager
python-dotenv
spotipy
pandas